feat(apps/api): implement isolation mode
