+++ .npmrc
@@ -1,4 +1,2 @@
 hoist = false
-recursive-install = false
 strict-peer-dependencies = false
-link-workspace-packages = false
