+++ apps/api/.npmrc
@@ -1,0 +1,1 @@
+recursive-install = false
