+++ apps/api/src/controllers/flt/index.ts
@@ -3,6 +3,8 @@ import GetFuseLeveragedTokenBySymbol from "./GetFuseLeveragedTokenBySymbol";
 import GetFuseLeveragedTokenUserPositionById from "./GetFuseLeveragedTokenUserPositionById";
 import GetFuseLeveragedTokenSwapsBySymbol from "./GetFuseLeveragedTokenSwapsBySymbol";
 import GetFuseLeveragedTokensGains from "./GetFuseLeveragedTokensGains";
+import GetFuseLeveragedTokenChartsBySymbol from "./GetFuseLeveragedTokenChartsBySymbol";
+import GetFuseLeveragedTokenBackingsBySymbol from "./GetFuseLeveragedTokenBackingsBySymbol";
 
 const controllers = {
     GetFuseLeveragedTokensByChainId,
@@ -10,6 +12,8 @@ const controllers = {
     GetFuseLeveragedTokenUserPositionById,
     GetFuseLeveragedTokenSwapsBySymbol,
     GetFuseLeveragedTokensGains,
+    GetFuseLeveragedTokenChartsBySymbol,
+    GetFuseLeveragedTokenBackingsBySymbol,
 };
 
 export default controllers;
