+++ apps/api/src/routers/index.ts
@@ -1,54 +1,48 @@
 import express, { Router } from "express";
-import fltsController from "../controllers/flts";
 import flt from "../controllers/flt";
+import ChainIdValidation from "../utils/validationChainId";
 
 const router: Router = express.Router();
 
 router.get(
     "/:chainId/flts",
-    // TODO(pyk): refactor this
-    fltsController.GetFuseLeveragedTokensByChainIdValidation,
+    ChainIdValidation,
     flt.GetFuseLeveragedTokensByChainId
 );
 
 router.get(
     "/:chainId/flts/:symbol",
-    // TODO(pyk): refactor this
-    fltsController.GetFuseLeveragedTokensByChainIdValidation,
+    ChainIdValidation,
     flt.GetFuseLeveragedTokenBySymbol
 );
 
-// TODO(pyk): refactor this
 router.get(
     "/:chainId/flts/:symbol/charts",
-    fltsController.GetFuseLeveragedTokensByChainIdValidation,
-    fltsController.GetFuseLeveragedTokenChartsBySymbol
+    ChainIdValidation,
+    flt.GetFuseLeveragedTokenChartsBySymbol
 );
 
 router.get(
     "/:chainId/flts/:symbol/swaps",
-    // TODO(pyk): refactor this
-    fltsController.GetFuseLeveragedTokensByChainIdValidation,
+    ChainIdValidation,
     flt.GetFuseLeveragedTokenSwapsBySymbol
 );
 
-// TODO(pyk): refactor this
 router.get(
     "/:chainId/flts/:symbol/backings",
-    fltsController.GetFuseLeveragedTokensByChainIdValidation,
-    fltsController.GetFuseLeveragedTokenBackingsBySymbol
+    ChainIdValidation,
+    flt.GetFuseLeveragedTokenBackingsBySymbol
 );
 
 router.get(
     "/:chainId/positions/:positionId",
-    // TODO(pyk): refactor this
-    fltsController.GetFuseLeveragedTokensByChainIdValidation,
+    ChainIdValidation,
     flt.GetFuseLeveragedTokenUserPositionById
 );
 
 router.get(
     "/:chainId/flts/insight/gains",
-    fltsController.GetFuseLeveragedTokensByChainIdValidation,
+    ChainIdValidation,
     flt.GetFuseLeveragedTokensGains
 );
